move populateWithContextDefaults functino to the pattern file
