+++ lib/pipeline-multi-env-gitops/index.ts
@@ -8,7 +8,29 @@ import { Construct } from 'constructs';
 // Team implementations
 import * as team from '../teams/pipeline-multi-env-gitops';
 
+export function populateWithContextDefaults(app: cdk.App, defaultAccount: string, defaultRegion: string) {
+    // Populate Context Defaults for the pipeline account
+    let pipeline_account = app.node.tryGetContext('pipeline_account');
+    pipeline_account ??= defaultAccount;
+    let pipeline_region = app.node.tryGetContext('pipeline_region');
+    pipeline_region ??= defaultRegion;
+    const pipelineEnv: cdk.Environment = { account: pipeline_account, region: pipeline_region };
 
+    // Populate Context Defaults for the Development account
+    let dev_account = app.node.tryGetContext('dev_account');
+    dev_account ??= defaultAccount;
+    let dev_region = app.node.tryGetContext('dev_region');
+    dev_region ??= defaultRegion;
+    const devEnv: cdk.Environment = { account: dev_account, region: dev_region };
+
+    // Populate Context Defaults for the Production  account
+    let prod_account = app.node.tryGetContext('prod_account');
+    prod_account ??= defaultAccount;
+    let prod_region = app.node.tryGetContext('prod_region');
+    prod_region ??= defaultRegion;
+    const prodEnv: cdk.Environment = { account: prod_account, region: prod_region };
+    return { devEnv, pipelineEnv, prodEnv };
+}
 
 export interface PipelineMultiEnvGitopsProps {
     /**
